scripts/requirements.txt
# Requirements for citation metrics scripts
# Install with: pip install -r scripts/requirements.txt

requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
scholarly>=1.7.0  # For Google Scholar (optional, may require proxies)
